+++ Minecraft.World/WorldGen/Layers/Layer.cpp
@@ -111,7 +111,7 @@ Layer::Layer(__int64 seedMixup)
 {
 	parent = nullptr;
 
-	// 4jcraft added all those calls abecause of signed int overflow
+	// 4jcraft added casts to prevent signed int overflow
 	this->seedMixup = seedMixup;
 	this->seedMixup *= (uint64_t) this->seedMixup * 6364136223846793005l + 1442695040888963407l;
 	this->seedMixup = (uint64_t) this->seedMixup + seedMixup;
@@ -125,7 +125,7 @@ void Layer::init(__int64 seed)
 {
 	this->seed = seed;
 	if (parent != NULL) parent->init(seed);
-	// 4jcraft added all those calls abecause of signed int overflow
+	// 4jcraft added casts to prevent signed int overflow
 	this->seed *= (uint64_t) this->seed * 6364136223846793005l + 1442695040888963407l;
 	this->seed = (uint64_t) this->seed + seedMixup;
 	this->seed *= (uint64_t) this->seed * 6364136223846793005l + 1442695040888963407l;
@@ -137,7 +137,7 @@ void Layer::init(__int64 seed)
 void Layer::initRandom(__int64 x, __int64 y)
 {
 	rval = seed;
-	// 4jcraft added all those calls abecause of signed int overflow
+	// 4jcraft added casts to prevent signed int overflow
 	rval *= (uint64_t) rval * 6364136223846793005l + 1442695040888963407l;
 	rval += (uint64_t) x;
 	rval *= (uint64_t) rval * 6364136223846793005l + 1442695040888963407l;
@@ -196,7 +196,8 @@ int Layer::nextRandom(int max)
 #endif
 
 	if (result < 0) result += max;
-	rval *= rval * 6364136223846793005l + 1442695040888963407l;
-	rval += seed;
+	// 4jcraft added cast to unsigned
+	rval *= (uint64_t) rval * 6364136223846793005l + 1442695040888963407l;
+	rval += (uint64_t) seed;
 	return result;
 }
